TUN address setting (AAA) (#923)
On mac and linux, adapter now has ability to set its own TUN address when it gets a Acquire ZPR Address Response. This means that you can bring up an adapter without specifying any ZPR address and things will just work.

If using external auth, the adapter will set an AAA address on TUN prior to opening a connection to the auth service.

This also moved the "noisy" logging of pings and echos to TRACE level without providing for any way to observe that. This will be fixed in #925 .

Requires visa service release 0.3.0